+++ src/Table/UI/views/columns/date.blade.php
@@ -1,5 +1,5 @@
 <span class="text-sm leading-5 text-grey-500">
     @foreach ($getItems() as $item)
-        {{ $item->getValue() }}
+        {{ $item->getValue($getLocale()) }}
     @endforeach
 </span>