+++ src/Table/UI/views/columns/image.blade.php
@@ -4,18 +4,18 @@
         @if ($item->hasLink())
             <a
                 href="{{ $item->getLink() }}"
-                title="{{ $item->getValue() }}"
+                title="{{ $item->getValue($getLocale()) }}"
                 {!! $item->shouldOpenInNewTab() ? 'target="_blank" rel="noopener"' : '' !!}
                 {!! $item->getCustomAttributesAsString() !!}
                 class="leading-5 text-grey-800 hover:underline hover:underline-offset-2"
             >
-                @if ($item->getValue())
-                    <img src="{{ $item->getValue() }}" class="h-full w-full object-cover" />
+                @if ($item->getValue($getLocale()))
+                    <img src="{{ $item->getValue($getLocale()) }}" class="h-full w-full object-cover" />
                 @endif
             </a>
         @else
-            @if ($item->getValue())
-                <img src="{{ $item->getValue() }}"
+            @if ($item->getValue($getLocale()))
+                <img src="{{ $item->getValue($getLocale()) }}"
                      class="h-full w-full object-cover" {!! $item->getCustomAttributesAsString() !!} />
             @else
                 <x-chief::icon.image class="size-5 text-grey-500" />