+++ src/Table/UI/views/columns/text.blade.php
@@ -2,16 +2,16 @@
     @if ($item->hasLink())
         <a
             href="{{ $item->getLink() }}"
-            title="{{ $item->getValue() }}"
+            title="{{ $item->getValue($getLocale()) }}"
             {!! $item->shouldOpenInNewTab() ? 'target="_blank" rel="noopener"' : '' !!}
             {!! $item->getCustomAttributesAsString() !!}
             class="leading-5 text-grey-800 hover:underline hover:underline-offset-2"
         >
-            {!! $item->getValue() !!}
+            {!! $item->getValue($getLocale()) !!}
         </a>
     @else
         <span class="leading-5 text-grey-500" {!! $item->getCustomAttributesAsString() !!}>
-            {!! $item->getValue() !!}
+            {!! $item->getValue($getLocale()) !!}
         </span>
     @endif
 @endforeach